// packages/disk-moon-city/moonCity/types.ts
export enum ATTACK_TYPE {
  SPY = 'spy',
  ATTACK_CITY = 'attack_city',
  FACTORY_SABOTAGE = 'factory_sabotage',
  POWER_STATION_SABOTAGE = 'power_station_sabotage',
  DESTROY_ENERGY_TRANSMITTER = 'destroy_energy_transmitter',
  DAMAGE_VEHICLE = 'damage_vehicle',
  WEAPON = 'weapon',
  VEHICLE_ATTACK = 'vehicle_attack'
}